+++ R/build_sources.R
@@ -265,11 +265,11 @@ cr_build_upload_gcs <- function(local,
   dir.create(deploy_folder, showWarnings = FALSE)
   myMessage(paste0("Copying files from ",
                    local, " to /", deploy_folder),
-            level = 3)
+            level = 2)
   file.copy(local, deploy_folder, recursive = TRUE)
   myMessage(paste0("Compressing files from /",
                    deploy_folder, " to ", tar_file),
-            level = 3)
+            level = 2)
   tar(tar_file,
       files = deploy_folder,
       compression = "gzip")